requirements.txt
 

django
django-ckeditor==6.7.2
django-cors-headers
celery
djangorestframework
Pillow
psycopg2-binary
djangorestframework-simplejwt
gunicorn
channels[daphne]
django-storages
boto3
load_dotenv
django-redis
python-freeDictionaryAPI
httpx
django-filter
django-celery-beat
markdown 
coreapi
# swagger:
django-rest-swagger
drf-yasg[validation]
pytest
 